24may26

+++ 2 семестр/новый автотест/task577/577.c
@@ -2,36 +2,44 @@
 #include <stdlib.h>
 #include <stdbool.h>
 
-int main(void) {
+int main(void)
+{
     int M, N, L, K;
     FILE *in = fopen("data.txt", "r");
     FILE *out = fopen("res.txt", "w");
 
-    if (in == NULL || out == NULL) {
+    if (in == NULL || out == NULL)
+    {
         if (out)
             fprintf(out, "ERROR\n");
-        if (in) fclose(in);
-        if (out) fclose(out);
+        if (in)
+            fclose(in);
+        if (out)
+            fclose(out);
         return -1;
     }
 
-    if (fscanf(in, "%d %d %d %d", &M, &N, &L, &K) != 4 || M <= 0 || L <= 0 || K <= 0) {
+    if (fscanf(in, "%d %d %d %d", &M, &N, &L, &K) != 4 || M <= 0 || L <= 0 || K <= 0)
+    {
         fprintf(out, "ERROR\n");
         fclose(in);
         fclose(out);
         return -1;
     }
 
-    int *A = (int*)malloc(L * K * sizeof(double));
-    if (!A) {
+    int *A = (int *)malloc(L * K * sizeof(double));
+    if (!A)
+    {
         fprintf(out, "ERROR\n");
         fclose(in);
         fclose(out);
         return -1;
     }
 
-    for (int i = 0; i < L * K; i++) {
-        if (fscanf(in, "%d", &A[i]) != 1) {
+    for (int i = 0; i < L * K; i++)
+    {
+        if (fscanf(in, "%d", &A[i]) != 1)
+        {
             fprintf(out, "ERROR\n");
             free(A);
             fclose(in);
@@ -43,7 +51,8 @@ int main(void) {
 
     // Флаги для строк и столбцов
     int *cols_with_mod = (int *)malloc(K * sizeof(int));
-    if (!cols_with_mod) {
+    if (!cols_with_mod)
+    {
         fprintf(out, "ERROR\n");
         free(A);
         free(cols_with_mod);
@@ -53,33 +62,43 @@ int main(void) {
 
     // Отметить строки и столбцы, содержащие элементы с остатком N по модулю M
     int count_good_cols = 0;
-    for (int i = 0; i < L; i++) {
-        for (int j = 0; j < K; j++) {
-            int val = (int)A[i * K + j];
-            if (val % M == N) {
+    for (int i = 0; i < L; i++)
+    {
+        for (int j = 0; j < K; j++)
+        {
+            int val = A[i * K + j];
+            if (val % M == N)
+            {
                 cols_with_mod[j] = 1;
             }
         }
     }
-    for (int j = 0; j < K; j++){
+    for (int j = 0; j < K; j++)
+    {
         if (cols_with_mod[j] == 1)
             count_good_cols++;
     }
 
     int *inds = (int *)malloc(count_good_cols * sizeof(int));
     int cur_ind = 0;
-    for (int j = 0; j < K; j++){
-        if (cols_with_mod[j]){
+    for (int j = 0; j < K; j++)
+    {
+        if (cols_with_mod[j])
+        {
             inds[cur_ind] = j;
             cur_ind++;
         }
     }
 
     // Модификация элементов по условиям задачи
-    for (int i = 0; i < L; i++) {
-        for (int iter = 0; iter < count_good_cols; iter++){
-            for (int j = 0; j < count_good_cols - iter - 1; j++){
-                if (A[i * K + inds[j]] > A[i * K + inds[j + 1]]){
+    for (int i = 0; i < L; i++)
+    {
+        for (int iter = 0; iter < count_good_cols; iter++)
+        {
+            for (int j = 0; j < count_good_cols - iter - 1; j++)
+            {
+                if (A[i * K + inds[j]] > A[i * K + inds[j + 1]])
+                {
                     int tmp = A[i * K + inds[j]];
                     A[i * K + inds[j]] = A[i * K + inds[j + 1]];
                     A[i * K + inds[j + 1]] = tmp;
@@ -90,8 +109,10 @@ int main(void) {
 
     // Вывод результатов
     fprintf(out, "%d %d %d %d\n", M, N, L, K);
-    for (int i = 0; i < L; i++) {
-        for (int j = 0; j < K; j++) {
+    for (int i = 0; i < L; i++)
+    {
+        for (int j = 0; j < K; j++)
+        {
             fprintf(out, "%d ", A[i * K + j]);
         }
         fprintf(out, "\n");
